fix: relative reusable workflow w/ HTTP (not HTTPS) github instance url (#1214)
One more attempt to fix this, in the vein of #1212 & #1213, as the end-to-end tests still don't pass due to their use of `http://` protocol for testing.  This approach maintains the consistency of having `RemoteReusableWorkflowWith...` always have a single format of data in its supplemental field, but changes it from a hostname to a base URL so that `gitHubInstance` can be `http://`.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1214): <!--number 1214 --><!--line 0 --><!--description Zml4OiByZWxhdGl2ZSByZXVzYWJsZSB3b3JrZmxvdyB3LyBIVFRQIChub3QgSFRUUFMpIGdpdGh1YiBpbnN0YW5jZSB1cmw=-->fix: relative reusable workflow w/ HTTP (not HTTPS) github instance url<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1214
Reviewed-by: limiting-factor <limiting-factor@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>