feat(jobparser): rewrite reusable workflow 'needs' references between jobs (#1217)
Reusable workflow expansion is a capability that will allow Forgejo to schedule every job of a reusable workflow as a separate job, allowing the use of multiple runners and clear independent logging.  This is a multi-PR implementation effort, project work tracking link: https://codeberg.org/forgejo/forgejo/issues/9768#issuecomment-8828016

This PR extends reusable workflow expansion such that if one job in a reusable workflow references another job via `${{ needs.some-job.outputs.some-output }}`, that expression is rewritten to reference the true ID of the job in the reusable workflow.  The expression would be mutated to `${{ needs[format('{0}.{1}', 'reusable-1', 'some-job')].outputs.some-output }}`, where `'reusable-1'` is the job ID of the parent/caller job.  With this change, jobs can reference each other, and are prevented from referencing other jobs in the workflow that aren't inside the scope of the reusable workflow.

In addition to test automation present here, this change has been manually tested on an [experimental Forgejo branch](https://codeberg.org/mfenniak/forgejo/pulls/3).

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1217): <!--number 1217 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiByZXdyaXRlIHJldXNhYmxlIHdvcmtmbG93ICduZWVkcycgcmVmZXJlbmNlcyBiZXR3ZWVuIGpvYnM=-->feat(jobparser): rewrite reusable workflow 'needs' references between jobs<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1217
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>