feat(jobparser): allow matrix expansion on caller of reusable workflows (#1220)
Reusable workflow expansion is a capability that will allow Forgejo to schedule every job of a reusable workflow as a separate job, allowing the use of multiple runners and clear independent logging.  This is a multi-PR implementation effort, project work tracking link: https://codeberg.org/forgejo/forgejo/issues/9768#issuecomment-8828016

This PR extends reusable workflow expansion so that the calling job can have `strategy.matrix` defined on it, and therefore generate multiple instances of the reusable workflow jobs.  This change is completed by teasing apart the matrix generation of the jobs and the final data preparation of the jobs, and performing the matrix generation before the reusable workflow expansion.

In addition to test automation present here, this change has been manually tested on an [experimental Forgejo branch](https://codeberg.org/mfenniak/forgejo/pulls/3).

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1220): <!--number 1220 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiBhbGxvdyBtYXRyaXggZXhwYW5zaW9uIG9uIGNhbGxlciBvZiByZXVzYWJsZSB3b3JrZmxvd3M=-->feat(jobparser): allow matrix expansion on caller of reusable workflows<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1220
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>