feat(jobparser): allow accessing `${{ needs... }}` when expanding reusable workflows (#1221)
Reusable workflow expansion is a capability that will allow Forgejo to schedule every job of a reusable workflow as a separate job, allowing the use of multiple runners and clear independent logging.  This is a multi-PR implementation effort, project work tracking link: https://codeberg.org/forgejo/forgejo/issues/9768#issuecomment-8828016

This PR adds the capability for expansion of reusable workflow jobs to be deferred until `${{ needs.... }` values can be evaluated.  This builds on existing support for doing this in simple jobs, where the jobs are returned to Forgejo and marked as "incomplete", and can later be passed back into the jobparser for a second-pass to fully evaluate them.  For reusable workflows, this deferral occurs in two places when accessing the `needs` context: from `strategy.matrix` and `with`.  Example dynamic resolution of reusable workflows:

```
on:
  pull_request:

jobs:
  pre-job:
    runs-on: ubuntu-slim
    outputs:
      scalar-value: ${{ steps.define.outputs.scalar }}
      array-value: ${{ steps.define.outputs.array }}
    steps:
      - id: define
        run: |
          echo 'scalar=scalar value' >> "$FORGEJO_OUTPUT"
          echo 'array=["value 1", "value 2"]' >> "$FORGEJO_OUTPUT"
  local-reusable-1:
    needs: pre-job
    uses: ./.github/workflows/reusable-1.yml
    strategy:
      matrix:
        dim: ${{ fromJSON(needs.pre-job.outputs.array-value) }}
    with:
      example-string-required: ${{ needs.pre-job.outputs.scalar-value }}
      context-github: ${{ matrix.dim }}
```

In addition to test automation present here, this change has been manually tested on an [experimental Forgejo branch](https://codeberg.org/mfenniak/forgejo/pulls/3).

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1221): <!--number 1221 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiBhbGxvdyBhY2Nlc3NpbmcgYCR7eyBuZWVkcy4uLiB9fWAgd2hlbiBleHBhbmRpbmcgcmV1c2FibGUgd29ya2Zsb3dz-->feat(jobparser): allow accessing `${{ needs... }}` when expanding reusable workflows<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1221
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>