fix(jobparser): track reusable workflow recursion limits across re-expansion (#1224)
Reusable workflow expansion is a capability that will allow Forgejo to schedule every job of a reusable workflow as a separate job, allowing the use of multiple runners and clear independent logging.  This is a multi-PR implementation effort, project work tracking link: https://codeberg.org/forgejo/forgejo/issues/9768#issuecomment-8828016

`jobparser` has to stop parsing a job when it hits a required input (`runs-on`, `strategy.matrix`, and `with`) that depends on the output of another job, and mark those jobs as incomplete for later revisiting when the dependency is complete.  This interrupts the internal state tracking that limits the recursion of reusable workflows to 5 levels.  This PR adds that recursion depth as a tracked outcome for incomplete jobs in reusable workflows, allowing the limit to be enforced across incomplete jobs when they are expanded.

Although marked as a bug, this is a bug/fixme/todo in unreleased functionality (#1209) that is not user impacting at this time.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1224): <!--number 1224 --><!--line 0 --><!--description Zml4KGpvYnBhcnNlcik6IHRyYWNrIHJldXNhYmxlIHdvcmtmbG93IHJlY3Vyc2lvbiBsaW1pdHMgYWNyb3NzIHJlLWV4cGFuc2lvbg==-->fix(jobparser): track reusable workflow recursion limits across re-expansion<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1224
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>