fix(jobparser): allow access to 'inputs' context when evaluating workflow call outputs (#1225)
Reusable workflow expansion is a capability that will allow Forgejo to schedule every job of a reusable workflow as a separate job, allowing the use of multiple runners and clear independent logging.  This is a multi-PR implementation effort, project work tracking link: https://codeberg.org/forgejo/forgejo/issues/9768#issuecomment-8828016

`inputs` is documented as a valid context when evaluating `on.workflow_call.outputs`.  For expanded reusable workflows, outputs are evaluated on the virtual "caller" job, and the inputs were not available previously.  This PR adds a stored copy of the inputs on the caller job, allowing this context to become available when later evaluating the outputs of the reusable workflow.

A minor documentation comment is also added in this PR noting that the reusable workflow runner is obsolete when reusable workflow expansion is being used, for a little clarity for future developers.

In addition to test automation present here, this change has been manually tested on an [experimental Forgejo branch](https://codeberg.org/mfenniak/forgejo/pulls/3).

Although marked as a bug, this is a bug/fixme/todo in unreleased functionality (#1216) that is not user impacting at this time.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1225): <!--number 1225 --><!--line 0 --><!--description Zml4KGpvYnBhcnNlcik6IGFsbG93IGFjY2VzcyB0byAnaW5wdXRzJyBjb250ZXh0IHdoZW4gZXZhbHVhdGluZyB3b3JrZmxvdyBjYWxsIG91dHB1dHM=-->fix(jobparser): allow access to 'inputs' context when evaluating workflow call outputs<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1225
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>