feat(jobparser): add tracking IDs for outer/inner jobs in reusable workflows (#1227)
When expanding reusable workflows, this PR adds new fields to the `SingleWorkflow` object allowing for tracking of which outer/parent jobs in a reusable workflow generated which inner jobs.  This was almost possible already by using job IDs, but it wouldn't take into account matrix evaluation which can generate multiple jobs with the same job ID.

This will enable:
- `jobs.<job_id>.secrets` will require knowing a precise parent/child relationship, both for `inherit` capability, but also for child reusable workflows to only be able to pass their secrets into another layer of reusable workflow
- simplifying detection in https://codeberg.org/forgejo/forgejo/pulls/10448 about whether a job is an outer workflow call, or inner workflow call
- future UI enhancements to display job hierarchy

Internal metadata that might actually be kept around in a runnable job is moved into an `__metadata` key so that it can be exempt from workflow schema validation in the near future.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1227): <!--number 1227 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiBhZGQgdHJhY2tpbmcgSURzIGZvciBvdXRlci9pbm5lciBqb2JzIGluIHJldXNhYmxlIHdvcmtmbG93cw==-->feat(jobparser): add tracking IDs for outer/inner jobs in reusable workflows<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1227
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>