feat(jobparser): expose API for `EvaluateWorkflowCallSecrets` (#1229)
When a reusable workflow is expanded into inner jobs by the jobparser, it will be necessary for Forgejo to evaluate `jobs.<job_id>.secrets` in order to provide secrets to the inner job, referencing the Forgejo encrypted secrets as well as other available job contexts.  This PR provides an API for Forgejo to evaluate those secrets after providing the necessary data to populate the required contexts.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1229): <!--number 1229 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiBleHBvc2UgQVBJIGZvciBgRXZhbHVhdGVXb3JrZmxvd0NhbGxTZWNyZXRzYA==-->feat(jobparser): expose API for `EvaluateWorkflowCallSecrets`<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1229
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>