feat(jobparser): expose Job to reusable workflow fetchers (#1234)
Add the parameter `job *jobparser.Job` to the reusable workflow fetchers.  This is intended to allow Forgejo to make a decision to opt-in or opt-out of the new reusable workflow expansion by examining the `runs-on` parameter of the job, and returning the `ErrUnsupportedReusableWorkflowFetch` error case if it doesn't want to expand the workflow.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1234): <!--number 1234 --><!--line 0 --><!--description ZmVhdChqb2JwYXJzZXIpOiBleHBvc2UgSm9iIHRvIHJldXNhYmxlIHdvcmtmbG93IGZldGNoZXJz-->feat(jobparser): expose Job to reusable workflow fetchers<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1234
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>