feat: request up to `capacity` jobs from Forgejo in one API call (#1245)
Permits the Runner to fetch multiple tasks from Forgejo in one API call, up to the configured capacity (minus any running jobs).

In the event that Forgejo doesn't support the `TaskCapacity` parameter (eg. current Forgejo), this implementation continues to perform `FetchTask` operations retrieving one task every configured `FetchInterval`, which is the same as the current behaviour.  The implementation is changed so that this is done in a single goroutine rather than one goroutine per capacity, but the behaviour is the same -- the interval was implemented through a shared rate limiter before, and now only has a single goroutine waiting on the same rate limiter.

[`testing/synctest`](https://pkg.go.dev/testing/synctest) is used to create neat unit tests for the new behaviour.  They look very fragile with millisecond precision waits, but are not due to synctest.

Related Forgejo PR: https://codeberg.org/forgejo/forgejo/pulls/10602

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1245): <!--number 1245 --><!--line 0 --><!--description ZmVhdDogcmVxdWVzdCB1cCB0byBgY2FwYWNpdHlgIGpvYnMgZnJvbSBGb3JnZWpvIGluIG9uZSBBUEkgY2FsbA==-->feat: request up to `capacity` jobs from Forgejo in one API call<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1245
Reviewed-by: Gusted <gusted@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>